+++ fleet_adapter_mqtt/fleet_adapter_mqtt/fleet_adapter.py
@@ -18,10 +18,11 @@ import yaml
 import nudged
 import time
 import threading
+import datetime
 
 import rclpy
 import rclpy.node
-from rclpy.parameter import Parameter
+from   rclpy.parameter import Parameter
 
 import rmf_adapter as adpt
 import rmf_adapter.vehicletraits as traits
@@ -36,6 +37,7 @@ from functools import partial
 
 from .RobotCommandHandle import RobotCommandHandle
 from .RobotClientAPI import RobotAPI
+from .RobotDelivery import DeliveryTask
 
 # ------------------------------------------------------------------------------
 # Helper functions
@@ -95,8 +97,17 @@ def initialize_fleet(config_yaml, nav_graph_path, node, use_sim_time, server_uri
 
     if not fleet_config['publish_fleet_state']:
         fleet_handle.fleet_state_publish_period(None)
+    else:
+        fleet_state_update_frequency = fleet_config['publish_fleet_state']
+        fleet_handle.fleet_state_publish_period(
+        datetime.timedelta(seconds=1.0/fleet_state_update_frequency))
+
+    task_capabilities_config = fleet_config['task_capabilities']
+
     # Account for battery drain
     drain_battery = fleet_config['account_for_battery_drain']
+    lane_merge_distance = fleet_config['lane_merge_distance']
+    waypoint_merge_distance = fleet_config['waypoint_merge_distance']
     recharge_threshold = fleet_config['recharge_threshold']
     recharge_soc = fleet_config['recharge_soc']
     finishing_request = fleet_config['task_capabilities']['finishing_request']
@@ -136,6 +147,19 @@ def initialize_fleet(config_yaml, nav_graph_path, node, use_sim_time, server_uri
 
     fleet_handle.accept_task_requests(
         partial(_task_request_check, task_capabilities))
+    
+    def _consider(description: dict):
+        confirm = adpt.fleet_update_handle.Confirmation()
+        confirm.accept()
+        return confirm
+
+    # Configure this fleet to perform action category
+    if 'action_categories' in task_capabilities_config:
+        for cat in task_capabilities_config['action_categories']:
+            node.get_logger().info(
+                f"Fleet [{fleet_name}] is configured"
+                f" to perform action of category [{cat}]")
+            fleet_handle.add_performable_action(cat, _consider)
 
     # Transforms
     rmf_coordinates = config_yaml['reference_coordinates']['rmf']
@@ -162,6 +186,36 @@ def initialize_fleet(config_yaml, nav_graph_path, node, use_sim_time, server_uri
         """Insert a RobotUpdateHandle."""
         cmd_handle.update_handle = update_handle
 
+        def _action_executor(category: str,
+                             description: dict,
+                             execution:
+                             adpt.robot_update_handle.ActionExecution):
+            with cmd_handle._lock:
+                if len(description) > 0 and\
+                        description in cmd_handle.graph.keys:
+                    cmd_handle.action_waypoint_index = \
+                        cmd_handle.find_waypoint(description).index
+                else:
+                    cmd_handle.action_waypoint_index = \
+                        cmd_handle.last_known_waypoint_index
+                cmd_handle.on_waypoint = None
+                cmd_handle.on_lane = None
+                cmd_handle.action_execution = execution
+        # Set the action_executioner for the robot
+        cmd_handle.update_handle.set_action_executor(_action_executor)
+        if ("max_delay" in cmd_handle.config.keys()):
+            max_delay = cmd_handle.config["max_delay"]
+            cmd_handle.node.get_logger().info(
+                f"Setting max delay to {max_delay}s")
+            cmd_handle.update_handle.set_maximum_delay(max_delay)
+        if (cmd_handle.charger_waypoint_index <
+                cmd_handle.graph.num_waypoints):
+            cmd_handle.update_handle.set_charger_waypoint(
+                cmd_handle.charger_waypoint_index)
+        else:
+            cmd_handle.node.get_logger().warn(
+                "Invalid waypoint supplied for charger. "
+                "Using default nearest charger in the map")
     # Initialize robot API for this fleet
     api = RobotAPI(
         fleet_config['fleet_manager']['broker'],
@@ -174,7 +228,9 @@ def initialize_fleet(config_yaml, nav_graph_path, node, use_sim_time, server_uri
         fleet_config['fleet_manager']['feedback_topic'],
         fleet_config['fleet_manager']['result_topic'],
         fleet_config['fleet_manager']['battery_topic'],
-        fleet_config['fleet_manager']['distance'])
+        fleet_config['fleet_manager']['distance'],
+        fleet_config['delivery']['dispenser_req'],
+        fleet_config['delivery']['ingestor_req'])
 
     # Initialize robots for this fleet
 
@@ -243,7 +299,8 @@ def initialize_fleet(config_yaml, nav_graph_path, node, use_sim_time, server_uri
                         update_frequency=rmf_config.get(
                             'robot_state_update_frequency', 1),
                         adapter=adapter,
-                        api=api)
+                        api=api,
+                        lane_merge_distance=lane_merge_distance,)
 
                     if robot.initialized:
                         robots[robot_name] = robot
@@ -271,7 +328,7 @@ def initialize_fleet(config_yaml, nav_graph_path, node, use_sim_time, server_uri
 
     add_robots = threading.Thread(target=_add_fleet_robots, args=())
     add_robots.start()
-    return adapter
+    return adapter, api
 
 
 # ------------------------------------------------------------------------------
@@ -318,16 +375,21 @@ def main(argv=sys.argv):
     else:
         server_uri = args.server_uri
 
-    adapter = initialize_fleet(
+    adapter, api = initialize_fleet(
         config_yaml,
         nav_graph_path,
         node,
         args.use_sim_time,
         server_uri)
 
+    dis_res_topic = config_yaml['rmf_fleet']['delivery']['dispenser_res']
+    ing_res_topic = config_yaml['rmf_fleet']['delivery']['ingestor_res']
+    #Init delivery
+    delivery_task = DeliveryTask("delivery_task", api, dis_res_topic, ing_res_topic)
     # Create executor for the command handle node
     rclpy_executor = rclpy.executors.SingleThreadedExecutor()
     rclpy_executor.add_node(node)
+    rclpy_executor.add_node(delivery_task)
 
     # Start the fleet adapter
     rclpy_executor.spin()